Addressed most comments.
